convert the script and tests into a package with a __main__ and a tests module
